+++ helpers/utils/src/time/Time.cpp
@@ -14,22 +14,22 @@ Time& Time::getElapsed() {
   return *this;
 }
 
-int64_t Time::toSeconds() const {
+size_t Time::toSeconds() const {
   return std::chrono::duration_cast<std::chrono::seconds>(
       _now - _elapsedLastCall).count();
 }
 
-int64_t Time::toMilliseconds() const {
+size_t Time::toMilliseconds() const {
   return std::chrono::duration_cast<std::chrono::milliseconds>(
       _now - _elapsedLastCall).count();
 }
 
-int64_t Time::toMicroseconds() const {
+size_t Time::toMicroseconds() const {
   return std::chrono::duration_cast<std::chrono::microseconds>(
       _now - _elapsedLastCall).count();
 }
 
-int64_t Time::toNanoseconds() const {
+size_t Time::toNanoseconds() const {
   return std::chrono::duration_cast<std::chrono::nanoseconds>(
       _now - _elapsedLastCall).count();
 }
